Updated esp32 code and updated flink infrastructure (#27)

+++ lib/Moisture/Moisture.h
@@ -4,7 +4,7 @@
 class Moisture {
     public:
     Moisture(){};
-    static uint16_t getMoisture();
+    static float getMoisture();
 };
 
 #endif
